requirements.txt
asgiref==3.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
defusedxml==0.7.1
Django==3.1.14
django-allauth==0.51.0
django-ckeditor==5.9.0
django-cors-headers==3.10.0
django-filter==2.4.0
django-guardian==2.4.0
django-js-asset==2.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.2
drf-nested-routers==0.93.4
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.3
inflection==0.5.1
oauthlib==3.2.2
packaging==24.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
# Database support
psycopg2-binary==2.9.1
dj-database-url==0.5.0
python-dotenv==1.2.1
